Add DIGENV support to shell
Digenv from lab1 should be included in lab2.
We need to know where the binary is located, so it needs to be exported
as an environment variable called "DIGENV_PATH".

Set it before executing. If not you will get a error message.
The path needs to be an absolut path pointing to a binary.